Move internal clients to client/internal
