+++ cartoframes/client/internal/__init__.py
@@ -1,0 +1,5 @@
+from .api_client import APIClient
+
+
+def create_client(creds, session=None):
+    return APIClient(creds, session)
