+++ cartoframes/client/sql_client.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from ..client import create_client
+from .internal import create_client
 
 
 class SQLClient(object):
@@ -11,14 +11,17 @@ class SQLClient(object):
     def __init__(self, credentials, session=None):
         self._client = create_client(credentials, session)
 
-    def query(self):
+    def query(self, query):
         """
         TODO
         """
-        pass
+        return self._client.execute_query(query)
 
-    def execute(self):
+    def execute(self, query):
         """
         TODO
         """
-        pass
+        return self._client.execute_long_running_query(query)
+
+
+
